make adaptive layout, fix eslint, catch errs, loading

+++ src/components/Modals/components/DeleteValue.tsx
@@ -1,9 +1,12 @@
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
 import { setCloseModal } from '../../../store/modalsSlice';
-import { useAppDispatch } from '../../../hooks/hooks';
+import { useAppDispatch, useAppSelector } from '../../../hooks/hooks';
 import { removePost } from '../../../store/postsSlice';
 import { removeAlbum } from '../../../store/albumsSlice';
+import ModalLoader from './ModalLoader/ModalLoader';
 
 type DeleteValueProps = {
   id: number;
@@ -14,19 +17,37 @@ type DeleteValueProps = {
 const DeleteValue = ({ id, typeModal, textModal }: DeleteValueProps) => {
   const dispatch = useAppDispatch();
   const handleClose = () => dispatch(setCloseModal());
-  const handleRemove = () => {
-    switch (typeModal) {
-      case 'deletePost':
-        dispatch(removePost(id));
-        break;
-      case 'deleteAlbum':
-        dispatch(removeAlbum(id));
-        break;
-      default:
-        break;
+  const {
+    isLoadings: { removePostLoading },
+  } = useAppSelector((state) => state.postsSlice);
+  const {
+    isLoadings: { removeAlbumLoading },
+  } = useAppSelector((state) => state.albumsSlice);
+  const handleRemove = async () => {
+    if (typeModal === 'deletePost') {
+      try {
+        await dispatch(removePost(id));
+        toast.success('Пост успешно удален');
+      } catch {
+        toast.error('Ошибка при удалении поста');
+      }
     }
+
+    if (typeModal === 'deleteAlbum') {
+      try {
+        await dispatch(removeAlbum(id));
+        toast.success('Альбом успешно удален');
+      } catch (err: any) {
+        toast.error('Ошибка при удалении альбома');
+      }
+    }
+
     dispatch(setCloseModal());
   };
+
+  if (removeAlbumLoading || removePostLoading) {
+    return <ModalLoader />;
+  }
 
   return (
     <Modal show onHide={handleClose} centered>
